Update recursion pattern logic and pascal calculation
Agent-Logs-Url: https://github.com/Ammar-Yasser8/Algorithms/sessions/498f0df7-e3ba-4fd2-be51-85b3c1f59592

+++ Pattern-Problems/Logic-Challenges/Program.cs
@@ -10,7 +10,7 @@ for (int i = 0; i < rows; i++)
 {
     Console.Write(new string(' ', (rows - i) * 2));
 
-    int value = 1;
+    long value = 1;
     for (int j = 0; j <= i; j++)
     {
         Console.Write($"{value} ");
